Refactor a couple more functions into constituency/utils
